pyproject.toml
[project]
name = "uni-admission-crawler"
version = "0.1.0"
description = "Daily crawler for university admission updates."
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "tenacity>=8.2",
  "beautifulsoup4>=4.12",
  "lxml>=5.0",
  "feedparser>=6.0",
  "google-genai>=0.3",
  "pyyaml>=6.0",
  "jinja2>=3.1",
  "python-dateutil>=2.9",
  "structlog>=24.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-httpx>=0.30", "responses>=0.25", "ruff>=0.6"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
  "live: tests that call real external services and require configured API keys",
]

[tool.ruff]
line-length = 120
target-version = "py311"
